src/main/java/com/testcode/testcode/service/StudyService.java
package com.testcode.testcode.service;

import com.testcode.testcode.domain.Member;
import com.testcode.testcode.domain.Study;
import com.testcode.testcode.repository.StudyRepository;

import java.util.Optional;

public class StudyService {

    private final MemberService memberService;
    private final StudyRepository studyRepository;

    public StudyService(MemberService memberService, StudyRepository studyRepository) {
        assert memberService != null;
        assert studyRepository != null;

        this.memberService = memberService;
        this.studyRepository = studyRepository;
    }

    public Study createNewStudy(Long memberId,  Study study) {
        Optional<Member> member = memberService.findById(memberId);
        study.setOwner(member.orElseThrow(() ->  new IllegalArgumentException("MEMBER IS NULL")));
        Study newStudy = studyRepository.save(study);
        memberService.notify(newStudy);
        memberService.notify(member.get());
        return newStudy;
    }

    public Study openStudy(Study study) {
        study.open();
        Study openedStudy = studyRepository.save(study);
        memberService.notify(openedStudy);
        return  openedStudy;
    }
}
